requirements.txt
PyQt5>=5.15.0
faster-whisper>=1.0.0
openai-whisper>=20240930
torch>=2.0.0
transformers>=4.30.0
opencc>=1.1.0
ffmpeg-python>=0.2.0